# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Comment management service for Minisource",
        "title": "Comment Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:5010",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/comments": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comments"
                ],
                "summary": "List comments",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Resource type",
                        "name": "resource_type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Resource ID",
                        "name": "resource_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Status filter",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort field",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort order",
                        "name": "sort_order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_minisource_comment_internal_models.ListCommentsResponse"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comments"
                ],
                "summary": "Create a new comment",
                "parameters": [
                    {
                        "description": "Comment data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_minisource_comment_internal_models.CreateCommentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_minisource_comment_internal_models.Comment"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/comments/search": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comments"
                ],
                "summary": "Search comments",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search query",
                        "name": "q",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_minisource_comment_internal_models.Comment"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/comments/stats": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comments"
                ],
                "summary": "Get comment statistics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Resource type",
                        "name": "resource_type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Resource ID",
                        "name": "resource_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_minisource_comment_internal_models.CommentStats"
                        }
                    }
                }
            }
        },
        "/api/v1/comments/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comments"
                ],
                "summary": "Get a comment by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Comment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_minisource_comment_internal_models.Comment"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comments"
                ],
                "summary": "Update a comment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Comment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_minisource_comment_internal_models.UpdateCommentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_minisource_comment_internal_models.Comment"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comments"
                ],
                "summary": "Delete a comment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Comment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/comments/{id}/replies": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comments"
                ],
                "summary": "Get replies to a comment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Parent Comment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_minisource_comment_internal_models.Comment"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_minisource_comment_internal_models.Attachment": {
            "type": "object",
            "properties": {
                "filename": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "mimeType": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                },
                "type": {
                    "description": "image, video, file",
                    "type": "string"
                },
                "uploadedAt": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "github_com_minisource_comment_internal_models.Comment": {
            "type": "object",
            "properties": {
                "attachments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_minisource_comment_internal_models.Attachment"
                    }
                },
                "authorAvatar": {
                    "type": "string"
                },
                "authorEmail": {
                    "type": "string"
                },
                "authorId": {
                    "description": "Author info",
                    "type": "string"
                },
                "authorName": {
                    "type": "string"
                },
                "content": {
                    "description": "Content",
                    "type": "string"
                },
                "contentHtml": {
                    "description": "Sanitized HTML",
                    "type": "string"
                },
                "createdAt": {
                    "description": "Timestamps",
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "deletedBy": {
                    "type": "string"
                },
                "depth": {
                    "description": "Depth for nested replies",
                    "type": "integer"
                },
                "dislikeCount": {
                    "type": "integer"
                },
                "editHistory": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_minisource_comment_internal_models.EditRecord"
                    }
                },
                "flaggedWords": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "isAnonymous": {
                    "type": "boolean"
                },
                "isDeleted": {
                    "type": "boolean"
                },
                "isEdited": {
                    "type": "boolean"
                },
                "isPinned": {
                    "description": "Features",
                    "type": "boolean"
                },
                "likeCount": {
                    "type": "integer"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "moderatedAt": {
                    "type": "string"
                },
                "moderatedBy": {
                    "type": "string"
                },
                "parentId": {
                    "description": "For replies",
                    "type": "string"
                },
                "pinnedAt": {
                    "type": "string"
                },
                "pinnedBy": {
                    "type": "string"
                },
                "reactionCounts": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "rejectionReason": {
                    "type": "string"
                },
                "replyCount": {
                    "description": "Stats",
                    "type": "integer"
                },
                "reportCount": {
                    "type": "integer"
                },
                "resourceId": {
                    "description": "ID of the resource being commented on",
                    "type": "string"
                },
                "resourceType": {
                    "description": "e.g., \"product\", \"ticket\", \"article\"",
                    "type": "string"
                },
                "rootId": {
                    "description": "Root comment ID for nested replies",
                    "type": "string"
                },
                "status": {
                    "description": "Moderation",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_minisource_comment_internal_models.CommentStatus"
                        }
                    ]
                },
                "tenantId": {
                    "description": "Multi-tenant support (e.g., \"shop\", \"ticket\", \"blog\")",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "github_com_minisource_comment_internal_models.CommentStats": {
            "type": "object",
            "properties": {
                "approvedCount": {
                    "type": "integer"
                },
                "averageRating": {
                    "type": "number"
                },
                "pendingCount": {
                    "type": "integer"
                },
                "reactionBreakdown": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer",
                        "format": "int64"
                    }
                },
                "rejectedCount": {
                    "type": "integer"
                },
                "totalComments": {
                    "type": "integer"
                },
                "totalReactions": {
                    "type": "integer"
                }
            }
        },
        "github_com_minisource_comment_internal_models.CommentStatus": {
            "type": "string",
            "enum": [
                "pending",
                "approved",
                "rejected",
                "spam"
            ],
            "x-enum-varnames": [
                "StatusPending",
                "StatusApproved",
                "StatusRejected",
                "StatusSpam"
            ]
        },
        "github_com_minisource_comment_internal_models.CreateCommentRequest": {
            "type": "object",
            "required": [
                "content",
                "resourceId",
                "resourceType",
                "tenantId"
            ],
            "properties": {
                "attachments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_minisource_comment_internal_models.Attachment"
                    }
                },
                "authorName": {
                    "type": "string"
                },
                "content": {
                    "type": "string",
                    "maxLength": 5000,
                    "minLength": 1
                },
                "isAnonymous": {
                    "type": "boolean"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "parentId": {
                    "type": "string"
                },
                "resourceId": {
                    "type": "string"
                },
                "resourceType": {
                    "type": "string"
                },
                "tenantId": {
                    "type": "string"
                }
            }
        },
        "github_com_minisource_comment_internal_models.EditRecord": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "editedAt": {
                    "type": "string"
                },
                "editedBy": {
                    "type": "string"
                }
            }
        },
        "github_com_minisource_comment_internal_models.ListCommentsResponse": {
            "type": "object",
            "properties": {
                "comments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_minisource_comment_internal_models.Comment"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_minisource_comment_internal_models.UpdateCommentRequest": {
            "type": "object",
            "required": [
                "content"
            ],
            "properties": {
                "attachments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_minisource_comment_internal_models.Attachment"
                    }
                },
                "content": {
                    "type": "string",
                    "maxLength": 5000,
                    "minLength": 1
                }
            }
        },
...
        "internal_handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}